+++ semitemplates/ts/jest.config.js
@@ -6,7 +6,7 @@ module.exports = {
   maxWorkers: 1,
   globals: {
     'ts-jest': {
-      isolatedModules: true
-    }
-  }
+      isolatedModules: true,
+    },
+  },
 };